requirements.txt
aiofiles==0.8.0
ajsonrpc==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
bottle==0.12.23
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.1
django-crispy-forms==1.14.0
django-taggit==3.0.0
h11==0.13.0
idna==3.3
ifaddr==0.2.0
marshmallow==3.17.1
packaging==21.3
Pillow==9.2.0
platformio==6.1.4
pyelftools==0.29
pyparsing==3.0.9
pyserial==3.5
requests==2.28.1
semantic-version==2.10.0
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.20.4
tabulate==0.8.10
urllib3==1.26.12
uvicorn==0.18.3
wsproto==1.1.0
zeroconf==0.39.0
